Add control over meshes loaded in background, change selected meshes behavior and add "patch blanket" surface parcellation support
Original neuroglancer is trying to fetch a mesh from meshes folder only for regions manually selected by the user. This commit brings new method `setMeshesToLoad` to specify an array of segment ids for which meshes are to be loaded regardless of user selection.
The behavior of perspective view regarding selected regions is changed when above method is used. Meshes for selected regions are displayed in full and only selected meshes are displayed when "Slice" checkbox is unchecked.

New option `surfaceParcellation?: boolean` in config for the case of "patch blanket" segmentation where each mesh is not a full volume mesh of a segment, but is just a part of cortical surface parcellation. In this case all of the meshes are displayed at all times and clipped when front octant is removed.