Add `restoreState` method to NehubaViewer
To be used instead of `viewer.ngviewer.state.restoreState` which should be private anyway.

At the moment this method clears the state of the neuroglancer viewer before restoring and sets crossSectionBackground value from config if not present in the state. More nehuba-specific actions performed on changing state will be added in the future.