Add touch support from google/neuroglancer@ebcd94ef4c4c79770dc2ab7866c3e6b7efbc2351
Rename `enableShiftDrag` to `enablePerspectiveDrag` since it now controls both dragging with shift+mouse and by two fingers on the touch screen.
Also `enablePerspectiveDrag` is not toggleable anymore. Was not really needed and implementation is much simpler without it, the whole duplicating action listener is gone and does not need to be maintained.